Adds streaming microservice
This adds the first of the microservices - the streaming service.
Currently just tracking the terms "savethe8th" and "repealthe8th"
although it's open to adding more in the future.

Had surprisingly many problems just trying to access the Tweet's
text; retweets and tweet truncating were the main issues causing
this.

We're also filtering out tweets that aren't in English and Tweets
that originate outside the Irish timezone - this one should be
looked into more later