Adds classifier server and connects up streamer-ms
Despite being able to run the flask server as multiple workers
(using gunicorn) the pipeline is still effectively single threaded
due to the streamer. It waits for the classifier to respond before
getting the next Tweet. We need to find a way to make the streamer's
http posts async but this is a start for now.