Excludes tweets with either 0 or >1 relevant hashtags
Removing tweets with 0 hashtags because we don't want them in the
first place. This seems to happen if someone's replying to a
tweet containing the hashtag but they don't put it in their tweet.

Removing tweets with more than 1 hashtag because it means they said
both repeal and save so we don't know who to allocate it to. In
future if we start following more hashtags we might rethink this
decision but for now it's best to leave them out.