Adds initial anaylsis capabilities to the project
Currently it just generates static images using matplotlib to give us
an idea of what the data is looking like so far. I'll probably be using
scripts like these to generate once off graphs for the FYP report that
won't be displayed on the webpage.

For graphs that will be displayed on the web, the plan is to build an
analysis microservice from this and cache the aggregrated data.
Graphs can then be generated from this using javascript on the client's
machine resulting in a much more responsive graph than those generated
currently.