Replaces tweets with relaxed chats for wordcloud benchmark
This deals with the issues highlighted previously in that the
twitter corpus provided by the NLTK library is too politically
focused due to the data being collected during the UK general
election in 2015. This changes the benchmark to content of
conversations overheard and then transcribed. From looking
through the top words it's clear this is a much better dataset
to match everyday word usage.