Improves clarity of sentiment chart
Scales the sentiment range to [-1,1] to help clarify what the values
mean. Also added emojis in the y-axis label as an extra hint.

It does mean we lose some of the fancy animations as chartjs
animates from 0, which is now the middle of our graph. We don't
get it rising from the bottom like we used to. It's not a big
enough drawback to revert the change however so it's something
we'll have to deal with.