Creates config.js while building the web container
A bit of a hack but I think it's the most elegant solution
available to us at the moment. I'm inserting JS code to
the start of the config.json file which turns it into a
valid JS file. It does this by assigning the JSON blob
to a variable.

We can include this config.js file like any other JS file
and every JS file will be able to access its values. This
results in a far more efficent result than sending an
AJAX request for this JSON data on page load.